refactor(wds): internal hooks, utils move to internal directory
