+++ packages/wds/src/components/flex-box/style.ts
@@ -1,7 +1,7 @@
 import { css } from '@wanteddev/wds-engine';
 
 import { createResponsiveStyle } from '../../utils';
-import { toCssValue } from '../../utils/css';
+import { toCssValue } from '../../utils/internal/css';
 
 import type { Theme } from '@wanteddev/wds-engine';
 import type { FlexBoxDefaultProps, FlexBoxProps } from './types';
