+++ packages/wds/src/components/list/style.ts
@@ -6,7 +6,7 @@ import {
   getPreviousValue,
   typographyStyle,
 } from '../../utils';
-import { toCssValue } from '../../utils/css';
+import { toCssValue } from '../../utils/internal/css';
 
 import type { Theme } from '@wanteddev/wds-engine';
 import type { ListCellContentProps, ListCellProps } from './types';
