+++ packages/wds/src/components/search-field/style.ts
@@ -2,7 +2,7 @@ import { css } from '@wanteddev/wds-engine';
 
 import { typographyStyle } from '../../utils/typography';
 import { createResponsiveStyle } from '../../utils/responsive-props';
-import { toCssValue } from '../../utils/css';
+import { toCssValue } from '../../utils/internal/css';
 
 import type { SearchFieldProps } from './types';
 import type { SerializedStyles, Theme } from '@wanteddev/wds-engine';
