+++ packages/wds/src/components/segmented-control/index.tsx
@@ -14,8 +14,8 @@ import { Box } from '@wanteddev/wds-engine';
 import { usePrevious } from '@radix-ui/react-use-previous';
 
 import { FlexBox } from '../flex-box';
-import useResizeObserver from '../../hooks/use-resize-observer';
-import { calculateAnimationStyle } from '../../utils/animation';
+import useResizeObserver from '../../hooks/internal/use-resize-observer';
+import { calculateAnimationStyle } from '../../utils/internal/animation';
 import { VirtualCheckboxInput } from '../virtual-input';
 
 import {
