+++ packages/wds/src/components/skeleton/style.ts
@@ -2,7 +2,7 @@ import { css, getColorByToken, keyframes } from '@wanteddev/wds-engine';
 import objectPath from 'object-path';
 
 import { createResponsiveStyle } from '../../utils/responsive-props';
-import { toCssValue } from '../../utils/css';
+import { toCssValue } from '../../utils/internal/css';
 
 import type { SkeletonProps } from './types';
 import type { Theme } from '@wanteddev/wds-engine';
