+++ packages/wds/src/components/text-button/contexts.ts
@@ -1,4 +1,4 @@
-import createLooseContext from '../../hooks/use-loose-context';
+import createLooseContext from '../../hooks/internal/use-loose-context';
 
 import type { ThemeColorsToken } from '@wanteddev/wds-engine';
 import type { TextButtonVariant } from './types';
