+++ packages/wds/src/components/toggle-icon/style.ts
@@ -1,7 +1,7 @@
 import { css, getColorByToken } from '@wanteddev/wds-engine';
 
 import { createResponsiveStyle } from '../../utils/responsive-props';
-import { toCssValue } from '../../utils/css';
+import { toCssValue } from '../../utils/internal/css';
 
 import type { Theme } from '@wanteddev/wds-engine';
 import type { ToggleIconProps } from './types';
